REORG: reusable KEF code/tests in their own module,   so that encryption.py still has krux-specific code and   kef.py could be re-used by another project and/or cli
